add getTimelapse, which also helped in getTodos
